third_party/blink/web_tests/fast/scroll-snap/snaps-after-scrollbar-scrolling-tap.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1" />
<link rel="stylesheet" href="resources/simple-snap.css">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/gesture-util.js"></script>
<script src="../../resources/scrollbar-util.js"></script>
<script src="../../resources/blink-coordinates-util.js"></script>

<div id='scroller'>
  <div id="space"></div>
  <div class="snap left top"></div>
  <div class="snap right top"></div>
  <div class="snap left bottom"></div>
</div>

<script>
  internals.settings.setScrollAnimatorEnabled(false);

  window.onload = async () => {
    // MacOS devices do not have touch screens Tests are irrelevant.
    const macPlatform = navigator.platform.indexOf('Mac') == 0;
    if (shouldSkipTests(macPlatform,
                        'Platform does not support touch screens')) {
      return;
    }

    await waitForCompositorCommit();

    const scroller = document.getElementById("scroller");
    const snapPoint = 400;
    const trackPos = 350;

    function tapAndScroll(x, y) {
      const scrollend = waitForScrollendEvent(scroller);
      touchTapOn(x, y);
      return scrollend;
    }

    promise_test (async () => {
      await waitForScrollReset(scroller);
      const thumb = horizontalThumb(scroller);
      await tapAndScroll(trackPos, thumb.y);
      assert_approx_equals(scroller.scrollLeft, snapPoint, 1);
    }, "Snaps after tapping on the forward track on the horizontal scrollbar.");

    promise_test (async () => {
      await waitForScrollReset(scroller);
      const thumb = verticalThumb(scroller);
      await tapAndScroll(thumb.x, trackPos);
      assert_approx_equals(scroller.scrollTop, snapPoint, 1);
    }, "Snaps after tapping on the forward track on the vertical scrollbar.");
  };
</script>
